Making Region a volume property
The region must be consistent with the url property, so if specified "https://s3.eu-west-1.amazonaws.com" then the region must be "eu-west-1" otherwise an error is thrown.
So it makes no sense to have the region field inside the storage key values, as under a same volume they must be the same as specified in the url.

Moreover, when specifying the region, then the url is not required when working with amazon s3, as the endpoint resolver infers the url from the region itself.