Fix storage failure when region not specified.
The region MUST be specified in order to perform requests when using the aws-sdk-s3 crate as we do.
It is indeed necessary to communicate with the AWS S3 servers. However this is not the case when working
with MinIO, in which case it ignores the region.
In order to make it an optional parameter in the config file for these cases where we work with MinIO
we instead specify an empty region programmatically to the sdk config.